# public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"ConsoleDTO": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"manufacturer": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"manufacturer"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GameDTO": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"title": {
						"type": "string"
					},
					"console": {
						"$ref": "#/components/schemas/ConsoleDTO"
					}
				},
				"required": [
					"title"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ReviewDTO": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"game": {
						"$ref": "#/components/schemas/GameDTO"
					},
					"rating": {
						"type": "number",
						"format": "double"
					},
					"review_text": {
						"type": "string"
					}
				},
				"required": [
					"rating",
					"review_text"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "api-review-game",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/reviews": {
			"get": {
				"operationId": "GetAllReview",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ReviewDTO"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Reviews"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateReview",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ReviewDTO"
								}
							}
						}
					}
				},
				"tags": [
					"Reviews"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ReviewDTO"
							}
						}
					}
				}
			}
		},
		"/reviews/{id}": {
			"get": {
				"operationId": "GetReviewById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ReviewDTO"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Reviews"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateReview",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ReviewDTO"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Reviews"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ReviewDTO"
							}
						}
					}
				}
			}
		},
		"/games": {
			"get": {
				"operationId": "GetAllGames",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/GameDTO"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Games"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateGame",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GameDTO"
								}
							}
						}
					}
				},
				"tags": [
					"Games"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GameDTO"
							}
						}
					}
				}
			}
		},
		"/games/{id}": {
			"get": {
				"operationId": "GetGameById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/GameDTO"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Games"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateGame",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/GameDTO"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Games"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GameDTO"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteGame",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Games"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/consoles": {
			"get": {
				"operationId": "GetAllConsole",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ConsoleDTO"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Consoles"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateConsole",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ConsoleDTO"
								}
							}
						}
					}
				},
				"tags": [
					"Consoles"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ConsoleDTO"
							}
						}
					}
				}
			}
		},
		"/consoles/{id}": {
			"get": {
				"operationId": "GetConsoleById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ConsoleDTO"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Consoles"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteConsole",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Consoles"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateConsole",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ConsoleDTO"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Consoles"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ConsoleDTO"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}